break out consumer and producer from `pup.py`.
Previously, the kafka consumer and producer were in `pup.py`. This
commit moves them into ClientContext, a small class that also
maintains the producer queue.

The consumer's methods have been moved to `consumer.py`.

Finally, `get_commit_date` was moved to `get_version_info`, which
assembles the dict that mnm needs.
